move WorkspaceManagerResourceMonitorRecord related functionality to its own file
